Move feature engineering functions to utils.feature_engineering
